Added test vectors that fail
