+++ examples/rsa-pkcs1/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hacspec-rsa-pkcs22"
+name = "hacspec-rsa-pkcs1"
 version = "0.1.0"
 authors = ["Malte Thomsen <malteut@gmail.com>", "Marcus Rasmussen <marcusjr@live.dk>", "Tobias Vestergaard <tsv2907@gmail.com>"]
 edition = "2018"
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lib]
-path = "src/rsa-pkcs22.rs"
+path = "src/rsa-pkcs1.rs"
 
 [dependencies]
 hacspec-lib = { path = "../../lib" }
