+++ examples/typecheck_examples.sh
@@ -71,4 +71,4 @@ typecheck hacspec-bls12-381-hash    no-ec   no-fst    json      coq
 typecheck hacspec-sha512            no-ec   no-fst    json      coq
 typecheck hacspec-ed25519           no-ec   no-fst    json      coq
 typecheck hacspec-linalg            no-ec   no-fst    json      coq
-typecheck hacspec-rsa-pkcs22        no-ec   no-fst    json      coq
+typecheck hacspec-rsa-pkcs1         no-ec   no-fst    json      coq
